+++ internal/cli/util/register.go
@@ -16,5 +16,6 @@ func Register(rootCmd *cobra.Command) {
 	rootCmd.AddCommand(sauceCmd)
 	rootCmd.AddCommand(cleanCmd)
 	rootCmd.AddCommand(viewCmd)
+	rootCmd.AddCommand(dagCmd)
 	rootCmd.AddCommand(worktree.WorktreeCmd)
 }